test(install-script): refactor tests in prep for Windows
