+++ pkg/scripts_test/Makefile
@@ -4,6 +4,7 @@ ifneq ($(OS),windows)
 	GOTESTPREFIX ?= sudo env PATH="${PATH}"
 endif
 
+GOTEST=go test
 GOTESTBINARY=sumologic_scripts_tests.test
 
 # We build the test binary separately to avoid downloading modules as root
