+++ pkg/scripts_test/command_unix.go
