+++ pkg/scripts_test/common_linux.go
@@ -1,3 +1,5 @@
+// go:
+
 package sumologic_scripts_tests
 
 import (
